added XFA, built dist
added functionality in pdf.worker.js to get XFA data from PDFs. It does it by first reading each named entry in the xref array, checking the start and end points for that data, and then sub-binary-stringing out parts from the xfa data stream and plopping them into an object. in addition, i exposed this function through the API in api.js so that I could actually use it in production