got rid of dist-XFA, changed xfa acquisition logic
for some reason the import in gatsby project works if I build in dist, but not if I build in dist-XFA. I know it's sloppy git etiquette, but for now I just made a copy of my gulpfile installing to dist-xfa for later examination.

Also, through some basic testing (<script> tagging the built code and running some client-side pdf investigation), I found that my original method of accessing xfa (pdfManager.pdfDocument.xfa or something like that) was not working on the built code... xfa was undefined. Because of that, I looked around in the form and saw that xfa was sometimes pulled using  acroForm.get('XFA'), so I set up an ensureCatalog promise tree on acroForm (not sure if that was necessary, but hey it works), and through that I got the xfa.